remove idle from highlightPanel
