+++ apps/fishing-map/features/map/map-state.atom.ts
@@ -4,3 +4,8 @@ export const mapIdleAtom = atom<boolean>({
   key: 'mapIdle',
   default: false,
 })
+
+export const mapReadyAtom = atom<boolean>({
+  key: 'mapReady',
+  default: false,
+})
