+++ apps/fishing-map/features/map/map-state.hooks.ts
@@ -1,8 +1,7 @@
 import { useEffect } from 'react'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 import useMapInstance from 'features/map/map-context.hooks'
-import { useMapLoaded, useMapStyle } from 'features/map/map-style.hooks'
-import { mapIdleAtom } from 'features/map/map-features.atom'
+import { mapIdleAtom, mapReadyAtom } from 'features/map/map-state.atom'
 
 export const useSetMapIdleAtom = () => {
   // Used it once in Map.tsx the listeners only once
@@ -33,4 +32,20 @@ export const useSetMapIdleAtom = () => {
 export const useMapIdle = () => {
   const idle = useRecoilValue(mapIdleAtom)
   return idle
+}
+
+export const useMapReady = () => {
+  const loaded = useRecoilValue(mapReadyAtom)
+  return loaded
+}
+
+export const useMapLoaded = () => {
+  const idle = useRecoilValue(mapIdleAtom)
+  const map = useMapInstance()
+  const mapInstanceReady = map !== null
+  const mapFirstLoad = (map as any)?._loaded || false
+  const mapStyleLoad = map?.isStyleLoaded() || false
+  const areTilesLoaded = map?.areTilesLoaded() || false
+  const loaded = mapInstanceReady && mapFirstLoad && (idle || areTilesLoaded || mapStyleLoad)
+  return loaded
 }
